+++ batchflow/models/torch/resnet.py
@@ -183,7 +183,7 @@ class SEResNet18(ResNet18):
     @classmethod
     def default_config(cls):
         config = super().default_config()
-        config['body/encoder/blocks/se'] = True
+        config['body/encoder/blocks/attention'] = 'se'
         config['body/encoder/blocks/ratio'] = 16
         return config
 
@@ -192,7 +192,7 @@ class SEResNet34(ResNet34):
     @classmethod
     def default_config(cls):
         config = super().default_config()
-        config['body/encoder/blocks/se'] = True
+        config['body/encoder/blocks/attention'] = 'se'
         config['body/encoder/blocks/ratio'] = 16
         return config
 
@@ -201,7 +201,7 @@ class SEResNet50(ResNet50):
     @classmethod
     def default_config(cls):
         config = super().default_config()
-        config['body/encoder/blocks/se'] = True
+        config['body/encoder/blocks/attention'] = 'se'
         config['body/encoder/blocks/ratio'] = 16
         return config
 
@@ -210,7 +210,7 @@ class SEResNet101(ResNet101):
     @classmethod
     def default_config(cls):
         config = super().default_config()
-        config['body/encoder/blocks/se'] = True
+        config['body/encoder/blocks/attention'] = 'se'
         config['body/encoder/blocks/ratio'] = 16
         return config
 
@@ -219,7 +219,7 @@ class SEResNet152(ResNet152):
     @classmethod
     def default_config(cls):
         config = super().default_config()
-        config['body/encoder/blocks/se'] = True
+        config['body/encoder/blocks/attention'] = 'se'
         config['body/encoder/blocks/ratio'] = 16
         return config
 
@@ -230,7 +230,7 @@ class SEResNeXt18(ResNeXt18):
     @classmethod
     def default_config(cls):
         config = super().default_config()
-        config['body/encoder/blocks/se'] = True
+        config['body/encoder/blocks/attention'] = 'se'
         config['body/encoder/blocks/ratio'] = 16
         return config
 
@@ -239,7 +239,7 @@ class SEResNeXt34(ResNeXt34):
     @classmethod
     def default_config(cls):
         config = super().default_config()
-        config['body/encoder/blocks/se'] = True
+        config['body/encoder/blocks/attention'] = 'se'
         config['body/encoder/blocks/ratio'] = 16
         return config
 
@@ -248,7 +248,7 @@ class SEResNeXt50(ResNeXt50):
     @classmethod
     def default_config(cls):
         config = super().default_config()
-        config['body/encoder/blocks/se'] = True
+        config['body/encoder/blocks/attention'] = 'se'
         config['body/encoder/blocks/ratio'] = 16
         return config
 
@@ -257,7 +257,7 @@ class SEResNeXt101(ResNeXt101):
     @classmethod
     def default_config(cls):
         config = super().default_config()
-        config['body/encoder/blocks/se'] = True
+        config['body/encoder/blocks/attention'] = 'se'
         config['body/encoder/blocks/ratio'] = 16
         return config
 
@@ -266,6 +266,6 @@ class SEResNeXt152(ResNeXt152):
     @classmethod
     def default_config(cls):
         config = super().default_config()
-        config['body/encoder/blocks/se'] = True
+        config['body/encoder/blocks/attention'] = 'se'
         config['body/encoder/blocks/ratio'] = 16
         return config
